Let Plot components size themselves
